Moved host events into their own package
