+++ bumblebee/host/events/__init__.py
@@ -1,0 +1,4 @@
+from bumblebee.host.events.auth_flow_events import AuthFlowEvents
+from bumblebee.host.events.bot_events import BotEvents
+from bumblebee.host.events.host_events import HostEvents
+from bumblebee.host.events.job_events import JobEvents
