+++ bumblebee/host/events/host_events.py
@@ -1,0 +1,9 @@
+from bumblebee.host.framework.events import EventBag, Event
+
+
+class HostEvents(EventBag):
+    class Startup(Event):
+        pass
+
+    class Shutdown(Event):
+        pass
